backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
requests==2.32.3
numpy==1.26.4

# ESRGAN upscaling (Phase 1)
# Using Git version of basicsr for PyTorch 2.5 compatibility
git+https://github.com/XPixelGroup/BasicSR.git
realesrgan==0.3.0

# SDXL support (Phase 2)
diffusers==0.31.0
transformers==4.46.0
accelerate==1.2.0
safetensors==0.4.5

# Qwen GGUF multimodal (Phase 3)
llama-cpp-python==0.3.2
  # Note: Use latest version with multimodal support

# GGUF Diffusion Model Support
gguf>=0.10.0




# SwinIR dependency
timm==1.0.11


# SupResDiffGAN dependencies
omegaconf==2.3.0
lightning==2.2.2
einops==0.8.0
wandb==0.18.6
scipy==1.13.1
matplotlib==3.8.4
pandas==2.2.2
scikit-learn==1.4.2
tqdm==4.66.2
tensorboard==2.16.2
seaborn==0.13.1
scikit-image==0.24.0
hydra-core==1.3.2
opencv-python>=4.0
